Use correct redirection code
HTTP 301 is "moved permanently" -- it can and will be cached by your
browser. In particular, this will cause the Wiki to always redirect to
/MainPage/edit when it is first built and run, since Firefox is caching
that initial redirect. The correct redirection code to use depends upon
the specific case of the application (and might even be 301), but
usually you want 302 "found" instead.