Fix sqlite double and text parameter binding
Both text and double types were being bound as blobs. This happened to
work in the case of text, since it came out as it went in (e.g. in
"WHERE foo = bar", bar would be bound to a blob, which would match the
blob that was inserted in the first place). In the case of double, the
C interface wasn't even being used correctly -- the only thing saving
a segfault was that this function will never be emitted by squall.
(Note that I still haven't tested this function, but it at least
*looks* right now.)