Add curl MLton FFI bindings
Getting the data at a URL can be a useful thing to do. The best way to
do it is probably with libcurl bindings. This unfortunately requires a
small C wrapper to do some bookkeeping (curl only wants to write via a
function pointer or a FILE*) but it's still pretty braindead.